refactor(ts): migrate Selector component
