+++ src/components/Selector/__tests__/Selector-test.tsx
@@ -3,6 +3,7 @@
 import { h } from 'preact';
 import Selector from '../Selector';
 import { mount } from 'enzyme';
+import { ReactElementLike } from 'prop-types';
 
 describe('Selector', () => {
   it('should render <Selector/> with strings', () => {
@@ -19,7 +20,7 @@ describe('Selector', () => {
         { value: 'index-b', label: 'Index B' },
       ],
     };
-    const tree = mount(<Selector {...props} />);
+    const tree = mount((<Selector {...props} />) as ReactElementLike);
 
     expect(tree).toMatchSnapshot();
   });
@@ -38,7 +39,7 @@ describe('Selector', () => {
         { value: 20, label: '20 results per page' },
       ],
     };
-    const tree = mount(<Selector {...props} />);
+    const tree = mount((<Selector {...props} />) as ReactElementLike);
 
     expect(tree).toMatchSnapshot();
   });
